rename NetworkView component to NetworkRenderer
