+++ src/components/NetworkRenderer.tsx
